Dev: app.js 및 nav.jsx 코드 구조 변경
